requirements.txt
# Flask and API
flask>=3.0.0

# TTS Dependencies
transformers>=4.30.0
torch>=2.0.0
soundfile>=0.12.1

# Twilio for phone calls
twilio>=8.0.0

# Environment variables
python-dotenv>=1.0.0

# Optional: For production deployment
gunicorn>=21.0.0

